+++ t/CodeLayout/ProhibitLongLines/basic.t
@@ -10,10 +10,10 @@ use experimental qw( signatures );
 
 # Test the policy directly without using Perl::Critic framework
 use lib                                               qw( lib t/lib );
-use Perl::Critic::Policy::CodeLayout::LimitLineLength ();
+use Perl::Critic::Policy::CodeLayout::ProhibitLongLines ();
 use ViolationFinder qw( bad count_violations good );
 
-my $Policy = Perl::Critic::Policy::CodeLayout::LimitLineLength->new;
+my $Policy = Perl::Critic::Policy::CodeLayout::ProhibitLongLines->new;
 
 subtest "Policy methods" => sub {
   # Test default_themes